🔧 Fix ALL PyO3/clippy issues for CI/CD
- Fixed downcast() calls to include py parameter (PyO3 0.20 compat)
- Fixed .map_err closures for proper error handling
- Fixed calculate_pi return type (String not PiCalculationResult)
- Fixed needless borrows in xmss.rs and mceliece.rs
- Fixed manual abs_diff pattern in genesis.rs
- Fixed clippy warnings in cli calculate command
- Moved test module to end of file
- Fixed integration tests (removed assert!(true))
- All clippy issues resolved
- Build passes with only benign warnings

Ready for GitHub Actions CI/CD!

+++ crates/consensus/src/difficulty.rs
@@ -84,7 +84,7 @@ mod tests {
 
         let mut blocks = Vec::new();
         for i in 0..2016 {
-            let mut block = Block::new(Hash::zero(), i as u64);
+            let mut block = Block::new(Hash::zero(), i);
             block.header.timestamp = i * 30 * 1000;
             blocks.push(block);
         }
@@ -101,7 +101,7 @@ mod tests {
 
         let mut blocks = Vec::new();
         for i in 0..2016 {
-            let mut block = Block::new(Hash::zero(), i as u64);
+            let mut block = Block::new(Hash::zero(), i);
             block.header.timestamp = i * 20 * 1000;
             blocks.push(block);
         }
